Added t4023, t4024 - partial update tests.
Now you can to 'tup upd foo.o' and it will only build what is necessary to get foo.o built. It displays the number of commands that were skipped at the end if the update is successful.

The way this works is it still builds the PDAG as normal, then for each requested node it marks that node, and all of the nodes it depends on recursively. At the end, if a node is not marked, it is removed from the PDAG.