Add t4016, t4024, t4025 - chdir tests.
Now chdir() is supported (to a certain degree) in ldpreload. The path gets passed back to the server, so tup can keep track of the current working directory when files are accessed.

Primarily this is needed to allow 'git describe' to be used in build processes to generate a build ID, like in tup itself.