pel_group: Use tup_entry_find instead of tup_entry_add.
Although the tupid passed into append_path_elements from start_server is already going to be in the entry table anyway, it looks dangerous to potentially be doing a database query from the updater without the db lock. Actually taking the lock would be superfluous, but it's better to avoid the issue altogether by just using the find() function.