Replace '~' symbol with '^'
Tilde element is a source of confusion for bash/shell.

While some sources mention that tilde element should not be
expanded in POSIX mode [1], POSIX document itself [2] does
not mention such behavior.

POSIX shell should always expand it to $HOME of the user.
If such user does not exist then result is unspecified.

It is safer just to replace tilde symbol with something more
neutral, such as caret ('^').

[1] http://www.gnu.org/software/bash/manual/html_node/Bash-POSIX-Mode.html
[2] http://pubs.opengroup.org/onlinepubs/009695399/utilities/xcu_chap02.html#tag_02_06