Bring varsed back into tup proper.
I didn't like the idea of having 'tup' and 'tup-varsed' as separate programs, since it causes clutter. With the tup_send_event code pulled out of ldpreload, tup can call this directly. Now if tup is invoked as 'tup varsed', it will not go through the normal database initialization, but instead pull from the already opened var dictionary as the standalone client would.