Added t6046, t6047 - ghost->generated brokenness.
If a command tries to read from a file that is generated by another command, it will get an error unless that file is listed in the Tupfile. If the Tupfile tries to list a generated node before it is actually created by the Tupfile (ie: it is still in the delete list), that will also create an error. However, it is possible that the command will have a link from a ghost node, and then a later command will upgrade that ghost to a generated file. This bypasses the second check, and may end up inserting a circular dependency. Therefore, when upgrading to a generated file, we have to pretend to remove the ghost node from the DAG by putting all the dependencies into modify/create (so they are re-executed/parsed as necessary) and remove all the links. Then we can flip the node to generated, and anything that depended on the ghost will be re-parsed/re-executed as necessary. So if we are in the first case (reading from a file, but not specifying it in the Tupfile), we will find that out during the update phase.