win32: Ignore certain test cases
Many of these aren't supported due to symlinks, or funky things in the shell that don't work in CMD.EXE. A brief comment for the reason the test is skipped is given in the check_no_windows argument, or in a comment before the check. It is possible some of these could be addressed, but at least this will give an easy spot check to see if things are broken on windows or not.