Use GCC from MacPorts.
GCC from XCode is horribly outdated because of the crappy Apple's licensing policy.
If Tup wants to use newer GCC/C features (such as __thread) it needs at least GCC 4.5.
Clang (which is eventually will become a standard C compiler on MacOSX) does not
support __thread yet. Support of thread-local storage (together with
C++0x thread_local and C1X _Thread_local) is going to be added
to it some day, but it is not there yet.

So the only option is to use the latest GCC. And build the fat binary using 'lipo' tool.

Note: if you have 64bit MacOSX version you need to install macport's GCC
with '+universal' variant.