fuse server: Use a single FUSE filesystem.
Instead of each server thread getting its own mountpoint, we can run each subprocess in the same filesystem and just use the process group ID to distinguish between them. Since we need to use pgid anyway to limit access to the filesystem, it seems like a good identifier (unless some subprocess tries to set pgid as well). This greatly reduces the startup time when running with a large -j number.