Added t6048 - test for bad tmp file usage.
This changes the fuse server to use per-job virtual directories instead of process groups to identify the source of a file access. When a file is opened for write, a temporary file is created in .tup/tmp instead of the real file-name. Attempts to read/rename/whatever this file are redirected to the .tup/tmp file, but only for processes running in that virtual directory. In this way, multiple jobs running simultaneously can't influence each other (eg: by all trying to write to "tmp.txt" at the same time).

This should also allow variants in the future, since we could run 'gcc -c foo.c -o foo.o' and 'gcc -DDEBUG -c foo.c -o foo.o' at the same time using the same files, but in different virtual directories. Then the outputs could just be moved to debug-variant/ or normal-variant/ or whatever after the command completes.