Added t6049 - don't remove unspecified outputs.
If a command writes to a file it didn't declare as an output, we shouldn't then delete that file from the file-system. This originally made sense when we were writing directly to the output files, but now that we are writing to temporary files, we can do all the checking before actually moving the files over. This way existing files aren't clobbered.