Added t4037, t4038, t4039 - allow mkdir/rmdir for tmp directories.
Running strip on an archive causes it to create a temporary directory, unpack the archive in that directory, and then strip all the objects before re-creating it. We can allow this by just keeping track of which directories have been created so we can return success in getattr(). We don't actually verify the paths, so as it is currently tup may allow things that would actually fail on the command line. That can be addressed if it comes up in the future.