win32: Add fake mappings to update mtimes.
Before the mtime on a node would be updated while iterating through the write_files list, but now it's done only if a matching mapping is found (since the target file won't exist yet). However, win32 doesn't yet do automatic temporary files, so we add a fake mapping in order to update the mtimes properly. In the future if that is added to win32, some of this can be cleaned up.