Automatically unmount .tup/mnt if it exists
In case if tup crashed or killed with SIGKILL .tup/mnt folder
is left in the broken state.

With this change tup deletes .tup/mnt in case of successful finalization.
In case of crash - such folder still exists on the filesystem.
Next time tup is running it checks for .tup/mnt and if it exists tup tries to
unmount it.
If the folder was not mounted then umount(2) returns EINVAL that is ignored.
On MacOSX (with macfuse 2.2.0) unmount does what it has to, on Ubuntu 10.04
with fuse 2.8.1 umount returns EPERM error, saying that only root can unmount
it (this is weird as the mount point is owned by me). So in case of EPERM we
ask the user to unmount the folder manually.