win32: Still need to use dt in file_entry structure
The fstatat() in file_set_mtime() was done assuming the filename can be reached from tup_top_fd(), as is in the case in the fuse server. However, sometimes Windows still uses partial file paths (for example, in 'echo foo > bar'), the 'bar' path will not be a full path. Also it seems the function to convert this to a full path is not thread safe.