Fix MacOSX race conditions in the tests
Before this change random tests failed because they were not able to remove
root-owned index directory created by Spotlight indexer
(.tup/mnt/.Spotlight-V100).

It seems that it is a race-condition between short-living fuse filesystem
and Spotlight. Tup sets fs flag "nobrowse" that tells Spotlight not
to index fs. But in case if Spotlight checks statfs() on the directory
*after* fuse fs was unmounted then it does not receive "nobrowse" flag.

To prevent it we use a MacOSX trick - create .metadata_never_index file
in the root of the filesystem. This forces Spotlight to skip this folder.

I run all tests 40 times in a row and everything is fine.