t2033: Don't need a dependency on a directory.
We only need dependencies on ghost directories. If the directory actually exists, we don't need a dependency on the command. Explicitly requiring this can cause a problem if the Tupfile points to the directory that uses it, because changing the Tupfile would then run all commands that the directory points to.