Use the server to watch the parsing stage.
This will enable arbitrary external scripts to be executed to generate :-rules. The file accesses of these scripts need to be watched in order to determine when the directory must be re-parsed. Additionally, watching the file accesses will remove the need to track symlinks internally, since that is handled by the fuse layer.