Remove explicit symlink handling.
With the ldpreload file watcher, tup needed to keep track of symlinks in order to make sure that any process reading from a symlink would properly get dependencies on the actual file read, as well as any symlinks that may have been used. However, the implementation of it didn't work in all cases, as evidenced by t5057. Fortunately the fuse file watcher essentially handles symlinks automatically, since we get a file access for the symlink as well as the underlying file. We don't need to otherwise know about the symlinks, and it's much easier to just not keep track of them.