updater: Make generated->file conversion more obvious.
When a generated file is modified, and the rule to create that file is removed, the file type goes from generated to a normal file. This happens during the file deletion stage, but it was not obvious unless you were paying attention to the colors. This change makes a separate header for it, though since the files are grouped together with the generated files that will be deleted, they share a main progress bar.