Change tup_dep_no_exist() behavior in test scripts.
This check would succeed if the two nodes specified had no link, or if either node did not exist. However, it is better to explicitly check for the missing node to not exist. This way if a tup_dep_no_exist check has a typo, it won't silently pass.

As a consequence, the node_exists() function had to be changed to check for the '0' directory and return 0 for the tupid. This functionality was moved to find_dir_tupid().