Added t6053 - broken update with resurrected files.
If a file is used as an input, tup has a check to make sure it won't be deleted (ie: because the rule that created it is removed). If we are going to resurrect the file, however, we can still use that file as an input. The parser shouldn't fail in this case.