win32: Add open_notify API for parser stage.
Since we don't keep track of opened files manually in the parser anymore, we need to either run the server to monitor the parser, or keep track of files elsewhere. In win32 it's a pain to run the server here, so we just hook the open call (which had to be hooked already) to add the file accesses to the file_info list.