Disable dir_mutex during parsing.
Since parsing is single threaded and we have the fuse thread in the same process as the parser thread, we run into deadlock issues with the dir_mutex. For example, an openat() in the parser will get the dir mutex, then call open, which results in the fuse thread waking up, which may call openat() and request the dir_mutex. The fix is to just not lock the mutex, but still do the fchdir for the *at wrappers. Since all file accesses now use *at functions and don't rely on the current directory, this works both with and without the compat wrappers.