Added t6055 - broken partial updates.
Since fuse is checking output file writes, the outputs must be unlinked before running the command. However, if the graph is pruned before executing a command with multiple outputs, some of the output nodes would be potentially missing. These outputs wouldn't be unlinked, so the command would fail. The fix is to make sure all outputs of a command are not pruned.