parser: Use fslurp_null() instead of fslurp()
Since the parser uses strchr() and the like, it should have a nul character at the end of the string. Just using fslurp() causes some tests to randomly fail in OSX as the parser walks off the end of the string. Not sure why this hasn't been a problem in Linux.