fuse server: Allow OSX zombies to access fuse fs.
With the 'ar' command, for some reason the unlink call to remove a .fuse_hidden file comes from a zombie libtool process. Further, OSX does not return a valid pgid for a zombie, but instead returns -1. This causes the context check to fail, so the unlink event gets ignored. This patch just implements a work-around specific to OSX, and allows accesses to the fuse fs if the getpgid() call fails with ESRCH.