Added t2094, t2095 - make sure readdir() works correctly.
When wild-carding in a Tupfile, the parser doesn't look at the filesystem, but instead looks at the tup database to see which files should be used. For example, a "*.o" input should match all objects that will be created but don't exist on disk yet, and not match ones that have been removed from the Tupfile but still exist. This presents a problem in run-scripts, which may use the scripting equivalent of 'for i in *.c', since that uses readdir(). Since we wrap readdir() in fuse, we have to avoid the usual route of going to the disk, and instead use the tup database entries so it matches the behavior of Tupfile parsing.

Limitations:
 - This currently does not set the inode or other parameters in the entry - it just fills out the name.
 - It does not allow a readdir() on a sub-directory other than the current one. The sub-directory may not have been parsed yet, and it would need to be parsed before the results would be accurate. We can't immediately parse it like we do in a Tupfile, however, because we are stuck in the readdir() in the single fuse thread. To support this, tup may have to use a multi-threaded fuse, but the fuse server currently assumes a single thread.