t2093 - Change run-script to run from Tupfile's dir.
It is sometimes useful to have a common file, and do 'include $(FOO_ROOT)/foo.tup' at the end of a Tupfile. If foo.tup runs a script, that script may need to look at files in the Tupfile's directoy (eg: to expand *.c). Therefore the run-script should run in the Tupfile's directory, not the included file's directory. This means run-scripts from included files need to be prefixed with $(TUP_CWD).