Added t2089, t2090 - more .gitignore tests.
t2089 tests for a particularly annoying condition that came up when fuse started to watch the parser stage. The .gitignore file was being removed before parsing, so if the parsing failed, the .gitignore file would be gone. Running things like 'git status' would then show all the generated files for the failed directory, which is messy. This fix makes it so the .gitignore file is only removed right before it is generated again.

t2090 just checks to make sure the .gitignore file is removed if the directive is no longer present. This normally gets cleaned up along with the rest of the generated files.