Fix t2095, t2092 - run-script readdir() on subdirectories.
Run-scripts may not necessarily fail if they can't readdir() on a subdirectory, even though tup returns an error to them. In this case, it's better to have tup fail so the error message can be caught and addressed by the user. It is still allowed to have a script access a subdirectory per t2092, but it can't be wild-carded at the moment.