Added t2096 - make sure fuse doesn't conflict with parser.
The run-script parser would be reading line-by-line as it was scheduled in order to add :-rules to the database. Meanwhile, the run-script would also be running, and occasionally it might do a readdir(), which would also access the tup database. There is no really good way to synchronize this. Instead, I opted to go back to the style where the parser reads in the entire output of the run-script before adding :-rules to the database. This ensures that the db access is synchronized. This means the list of files is essentially frozen when the run-script starts, and is then updated after the parser can go and process all of the :-rules. Without this change, t2096 would occasionally segfault.