valgrind: Use fuse-compatible sim-hints.
Since the parser and fuse fs are two threads in the same process, valgrind can hang waiting for a syscall (like chdir or fstat) to return, which will never happen since it won't let the fuse fs run. The fuse-compatible sim-hints tell valgrind that those syscalls may block, when on an ordinary file-system they would not.