Modify t2087 - failed parsing shouldn't break gitignore.
If a gitignore file was created for one directory, but parsing later fails for another directory, then the tup entry for the .gitignore file is never committed to the db. The next time that directory is parsed, it will see an existing .gitignore file with no tup entry and think it is a manual gitignore file, and break. Instead tup should just always create a .gitignore file if the directive is present, and manual .gitignore files be damned.