fuse: Add locking to the file_info structure.
This is somewhat superfluous since fuse is single-threaded, which writes to the structure. Also it only writes while the sub-process is active, and then tup uses the structure to write out dependencies, also in a single thread. The two parts are not active at the same time, so the locking isn't really necessary. However, it does cleanup messages from helgrind, and will be useful in the future to support a multi-threaded fuse if that is desired.