Use a master fork process to spawn sub-processes.
The sub-processes seem to be conflicting with fuse, since the fuse_thread is running in the main process that is spawning all the children. These children inherit things like the fuse file descriptor, which may cause problems. However, I am unable to create a simple test case that adequately demonstrates this. The behavior seen by fuse when the issue hits is either a deadlock, or a segfault. In the segfault case, fuse reads only 48 bytes for a RELEASEDIR request, which is too small. Fuse then reads beyond the end of the 48 bytes and sees information for the previous valid RELEASEDIR request, and fuse_lib_releasedir() attempts to free the fuse_dh structure a second time. For more information, see the tup thread at: http://groups.google.com/group/tup-users/browse_thread/thread/5838b1c4db2f3014/d66507b370e98049

The master fork process is spawned as early as possible to avoid inheriting any other tup information. This process just waits for information on a socket regarding the directory and command to run. The return value is written back to the socket along with the uniquely identifying server ID. In the main tup process, a thread reads the socket and stores the return value in a table. The worker threads wait until this table is updated to get their specific return value. This could be done better to avoid the broadcast.

For debugging this issue, libfuse was compiled into tup, and fuse.ko was modified to add printk's. 'ulimit -c unlimited' also provided core dumps, and gdb could provide backtraces. Helgrind gave some leads, but did not seem to reveal the actual issue at hand. To reveal the problem, t6048 was run in a loop like so:

while ./test.sh t6048-yacc.sh ; do x=$((x+1)); echo $x; done

I'm just documenting this in case I have to try it again, since I seem to remember tracking down a similar issue before, but forgetting entirely what the result was.