Added t6057 - missing inputs shouldn't cause circular dependencies.
The description in t6057 goes into more detail for the problem. The solution is to keep track of which files we are reporting as missing input dependencies, and make sure we don't add links to those. We can't skip the link creation entirely if there is a missing input dependency due to (at least) t5063.