Added t4048 - check for a big command string.
This switches the master fork socket to SOCK_STREAM instead of SOCK_DGRAM. Some commands in gittup are very large (17k), so just using PATH_MAX for the command string is not feasible. It's also entirely possible that we will hit the limit of a Unix domain DGRAM socket. Better to just use a stream to support arbitrary length commands, and a lock for atomicity.