Added t5073 - error messages for missing inputs & cmd failure.
I ran into this while writing Tupfiles for marfitude. I had forgotten to add a library as an input dependency to a linker rule, and also forgot to put some necessary files in the library. The link failed because the files weren't in the library (ie: undefined reference errors), and then when I added those files to the library, I was surprised that the library wasn't re-built before the link step. This was because the input checking was only done during success. This change just adds the input checking even if the command fails, so you can fix those errors before trying to fix the actual command errors.