display: Show the number of active commands.
This occurs only during the update stage, since that's the only time when we have multiple jobs. In this case, the progress bar so far will be displayed, along with the number of active jobs. When the job is finished, it is still displayed with its progress bar and any output & error messages.