display: Show failed sub-processes in red.
This requires saving off the dependency-checking errors to a temporary file, so that we know before we print those error messages whether or not it should have a red or blue banner.

Note that not all errors are sent to the file - for example, if there is a database error, this just gets written to stdout. These errors are typically not associated with a specific sub-process (like an unspecified output error), but are rather indications of tup bugs. Therefore they should be displayed as soon as possible.