display: Unify stdout/stderr for sub-processes.
This makes it easier to see the order in which things happen and gets rid of the annoying "tup: stdout" messages, but also prevents re-directing just the stderr for sub-processes. Not sure if that is ever useful, though. All of the stdout+stderr for a sub-process is directed to tup's stdout if it succeeds, or to tup's stderr if it fails.