Use .ini-style options instead of 'tup config'
Now you can create a project-level, user-level, and/or system-level options files in order to specify default options for tup, as suggested on the mailing list. This changes the behavior of autoupdate slightly, since the monitor now only reads the value at startup rather than every time files are changed. Additionally, existing options stored in the original 'tup config' database table or the ~/.tupconfig file are not converted over automatically.