display: Start counting at 1.
Now the Active progress bar just shows 0/N for the first line, and then displays the percentage complete after that. The sub-processes increment the number complete before the display is shown, so if there are 3 jobs, when the first finishes it will show 1/3, then 2/3, and then 3/3. Instead of showing a blank 3/3 line in all green, the all-green success status is moved to the final [ tup ] display.