Added t6060 - another circular dependency fatal tup error.
This one is caused by reading from two generated files without sticky links. If one of the generated files is connected to the other, the transitive dependency detection would kick in and add the second file as a link. This could cause a circular dependency, resulting in a fatal tup error during the next update.