master_fork: Only shutdown if it's initialized.
The inotify monitor closes down the master_fork process upon starting up, since it is not needed by the monitor. However, if the monitor shuts down successfully due to an 'rm -rf' on the .tup directory, it will try to shut it down again in tup_cleanup, resulting in misleading error messages.