Updates to parser display.
This saves the parser error messages to a tmpfile, and displays them only after the whole Tupfile has been parsed. This way messages can more easily be displayed under the banner of the Tupfile that is being parsed, so you can see which Tupfile failed because of an error, and which Tupfile failed because a dependent Tupfile failed. The failed parsings are also now displayed in red. This also finally gets rid of the annoying "Already parsed..." messages.