Allow use of the system's libsqlite3.
This serves as an example of how to do such a thing using an @-variable. Ie: Creating a tup.config with:

CONFIG_TUP_USE_SYSTEM_SQLITE=y

Will cause tup to be built with the system's libsqlite3 library. If it is not set or set to any value other than 'y', then tup is built with the included sqlite version.