Run sub-processes in a chroot if tup has set uid.
This allows tup to get all dependencies, even if a sub-process reads from a file with an absolute path (eg: '/usr/bin/gcc'). Currently tup does not track such dependencies, but that can be added now that it sees those accesses.

Tup drops privileges as soon as it is able. The main tup process drops privileges immediately after creating the master_fork process. The master_fork keeps privileges until the forked process can bind-mount /dev and chroot() into the fuse file-system, and then privileges are dropped for the sub-process.