win32: Fix threading issues with output buffering.
We have to make sure that the temporary output file we create is only inherited by one process, so this happens inside the dir_mutex. The file is created with the inherit flag set, then the process is created, and then we remove the file before giving up the mutex.

Also use FILE_ATTRIBUTE_TEMPORARY instead of FILE_ATTRIBUTE_NORMAL, since we expect it to be deleted almost immediately.

It is tricky to get Windows to automatically delete the file, so instead we have a server_postexec() function that will be called to remove the file after the updater is done displaying the data.