inotify monitor: Move where the autoupdate pid is cleared.
There is a race condition in clearing the autoupdate pid in the inotify monitor. For example, if 'tup flush' runs but the autoupdate hasn't finished yet, then 'flush' will see that the pid is not -1 and release the lock. The monitor will see this event and then set the pid to -1. It is possible that the flush will run again and regain the lock before autoupdate actually has a chance to run. Instead, it is better to use a special command 'tup autoupdate' which will tell the autoupdate process to clear the flag when it is finished. This way we guarantee that we've had a chance to run before clearing it. Trying to do it in the monitor itself is too messy, and not necessary. The error condition can be triggered by running t7005 repeatedly, and is easily triggered by doing a 'sleep(5);' before the tup_init() in the autoupdate sub-process.