Parser version 3: Added export keyword.
This allows you to 'export FOO', so that all subsequent :-rules and run-scripts in the Tupfile will have the FOO variable in their environments. Additionally, the environment variable is added to the tup database so that the correct Tupfiles will be re-parsed and commands re-executed in the event that the value of FOO changes from one invocation of tup to the next.

By default, only PATH is exported. All other environment variables must be explicitly exported.