fuse: Tweak the parser readdir() implementation.
Instead of querying the database from the fuse thread, the parser stashes away a list of the files for the current directory so they can be put into readdir() if a run-script wildcards the current directory. The behavior should be the same, but it keeps fuse away from the database, and should be easier to prevent threading errors in helgrind.