Added t4060 - support exporting environment variables that don't exist.
If a variable does not exist in the user's environment, and the Tupfile says to export it, then it should also not exist in the sub-process (rather than get exported as ""). These are stored as NULL values in the database, whereas a variable that is merely empty would be stored as "FOO=".