display: Only show progress bar on the last line.
The progress bar is only really meaningful on the last line, not on every job line. Also, the job line does not need to show the total number of jobs (essentially this information is redundant on each line). The progress bar expands to fill up the width of the console.