Track command execution time in the mtime field.
Since the mtime field of a command isn't currently used for anything, it can be used to track its execution time. This is a different usage from files, which tracks the file's actual mtime/ctime. For commands, it is now the number of milliseconds it took to execute. This can be used in future updates to approximate how long an update will take to complete.