win32: Wrap tmpfile() to fix Windows issue.
The parser creates a tmpfile for each directory that it parses to store output. For a project with many directories, we seem to hit a max tmpfile limit in Windows (637 or so). This fix provides a new implementation for tmpfile() that stores the file in the .tup/tmp directory with a known name and unique ID. The file retains the delete-on-close semantics of tmpfile(), so the parser doesn't have to change.