updater: Wrap graph creation & pruning in an SQL transaction.
The fact that prune_graph() was not in a transaction is a bug, since it inserts pruned commands into the modify list. Syncing on each one of those is not desirable. Wrapping the graph creation in a transaction is also slightly beneficial, since SQLite does not have to get the lock on the database for each select statement. In gittup.org for example, a full graph build goes from 460ms to 300ms.