dllinject: Wrap msvcrt.dll functions.
We need to wrap at least fopen, rename, and remove in msvcrt.dll. In particular, rename() is used by MinGW's ar.exe to rename the temporary file to the final library. Without this, tup complains that a random temporary file was written to.

The rename call was apparent by using 'strace -f ar crs libfoo.a ok.o'. To find the exact dll, 'objdump -p ar.exe' provides a listing of the dlls and function names.