Report errors when cleaning out .tup/tmp
If there is a failure removing a previous temporary file, the user won't find out until we try to create a file with the same ID some time later. Instead it is useful to know what the actual problem is, so the user can try to manually clean it up.