Lock the display mutex before printing server errors.
The servers just print directly to stderr, which may get jumbled if many processes are all failing for the same reason. This at least helps keep the error messages straight.

Note that this is for code that isn't "supposed" to fail, such as the creation of a sub-process, or creating its temporary output file. That is why these go directly to stderr instead of to the temporary output file, which is meant for the sub-process's output.