Use $tupcp alias for tests with cp on Windows.
Tup isn't able to monitor existing cp/copy/xcopy implementations, so tup-cp was created. The tests still call cp directly, so for those tests that run on Windows we need to alias 'cp' to 'tup-cp'. This is done simply by using $tupcp in the test itself, which is appropriately set based on the platform.