Added t6061 - fix for ghosts in a dir that gets removed.
This was discovered by looking through the code, not from a real-world failure. Still, this causes a dependency to be dropped that shouldn't have been. Essentially the directory removal code was too over-zealous in removing ghost nodes. Instead, the directory should be converted to a ghost itself if there is a ghost inside it.