display: Fix displaying multiple final progress bars.
The show_progress() function was trying to clear the progress bar if sum==total. However, two threads could see this condition because 'sum' is updated in show_result(), which happens in a separate lock from show_progress(). The progress bar is already cleared by execute_graph(), so we don't need to do it in show_progress() anyway.