windows: Use a temporary file instead of UDP.
Using a local UDP socket on Windows does not block if the buffer is full, so we can lose dependency data. Using a temporary file seems to be the easiest solution since we can lock the file in case a process spawns a sub-process and both attempt to access files.