display: Use job times in the percentage complete.
The percent complete metric was based only on the number of jobs. So if there are 2 jobs to do, one that takes 1s and one that takes 9s, after the first job completes tup would report 50% complete. However if we have the previous runtimes, it makes sense to instead report 10% complete. If we don't have the previous runtimes, it defaults to the job count as before.