windows: Use CreateFile instead of fopen for deps file.
It seems that the standard file buffering is getting in the way of atomically writing dependencies within the lock. FlushFileBuffers() would seem to be the answer, but that causes random other files (like 'touch foo' saying that 'foo' can't be found, when it is being used to *create* foo in the first place). The CreateFile API can allow us to specify append access, which isn't very well documented but seems to allow atomic writes to the end of the file.