Added t6062 - check for command strings of different types.
On Windows, trying to run "sh -c 'blah'" will try to open "blah", and "blah.exe" as files, causing ghosts to be created. If the 'sh -c' wrapper is removed, tup will find the existing 'blah' string in the database and use it as a command. However, it's type is a ghost. At some point commands will need to be separated from files/ghosts/directories/etc, but right now they are intermingled. Better to report an error when trying to insert it than to incorrectly add a ghost as a command and cause all sorts of trouble.