Added t6063 - Make sure input dependencies are not commands.
Similar to t6062, where tup gets confused if a command string is used as a filename elsewhere. In this case, we create a command with a valid Tupfile, then try to create a new command which will end up trying to read from a filename that is the command string. Ideally commands would be in a separate namespace, but at least now there is an error message rather than inserting bad things into the DAG.