windows: Detect bash shell.
If bash gets wrapped in a cmd.exe invocation, somehow the single-quotes get messed up and included as part of the filename instead of delimiting command arguments. This is a poor-man's work-around to detect if bash is explicitly being called, until such time as the shell issues in Windows get a better solution.