Track dependencies outside the tup hieararchy.
Currently this only adds the deps to the database, it does not yet rebuild if these files change (since they aren't scanned at the moment). All files outside of the tup hierarchy are added as ghosts, even if they are valid files. This way they don't get put into the tree during the initial scan. We'll know if they are actual files or not based on whether or not mtime is set.