windows: Distinguish between a root path and path separator.
Previously, both a slash (or backslash) and a string like "C:" would be considered a path separator. However, for full path detection we need to keep the "C:" part around since we have to know which drive a file is on. This also means we can't treat everything as if it is rooted as "/", so there is some special logic for Windows to check if the pathname is "C:", and then not try to do fstatat/openat and the like.