display: Show absolute time at each tup step.
Although the individual step time might be nice to print on a line, it is difficult to do it in a way that isn't confusing as to what step the time corresponds to (ie: we want to print "Scanning filesystem..." before doing the scan, but we don't know how long it will take until it finishes). This at least gets the warnings & other messages during the scan on separate lines and doesn't interfere with the timestamp display. We also get an absolute time at each step, so not just the scan is quantified.