Added t4064 - make sure mtimes are updated on unreachable external files.
One issue with this is due to the way fuse accesses files - if we try to access '/tmp/foo/bar', but the 'tmp/foo' directory doesn't exist, then we don't ever see the attempt to access 'bar'. As a consequence, we have to keep track of when directories are added so that we can re-run commands that may now be able to access files that they were trying to access. This is done by setting the mtime of an external file to 0 for directories, -1 for non-existent files, and the actual mtime for regular files / symlinks.