Added t4066 - check for error message if we can't support full_deps.
In Linux/OSX we need to have suid so we can run in a chroot and mount the /dev file-system in order to get full dependencies. If we try to specify full_deps but can't support it, the user should get an error message. This moves the error message to be up front in the updater rather than when tup goes to run the actual sub-process.