Update correctly if full_deps is toggled.
If we are enabling full_deps, then we have to rebuild everything in order to pick up the new dependencies on external files. If instead we are disabling full_deps, then we can just wipe out the entire /-tree. In this case, nothing needs to be immediately rebuild since we already have all the local dependencies anyway.