Added t7047 - monitor shouldn't die on ephemeral files.
If we get a file modification, then the file is deleted while we are flushing the event log, the fstatat() will fail since the file no longer exists, and the monitor quits. Instead we should just assume that we will be getting a file deletion (or moved from) event soon anyway, and call tup_file_del().