windows: Remove check for files with /temp/ in the path.
Windows would ignore files with /temp/ anywhere in the path, which could cause it to ignore real files if a user happens to create a temporary directory for themselves. There doesn't appear to be anything in gcc or VC that breaks as a result of taking this out, and all test cases still pass. If we need to ignore some temporary files in the future, this should be done in a smarter way (maybe checking $TMP / $TEMP and comparing) rather than wholesale dismissing everything with /temp/.