db: Split ghost reclaimable query into two.
The ghost reclaimable check needs to see if there are any other nodes that have the ghost as a directory, as well as check if the ghost points to anybody. I tried to be clever and combine this into one query, but unfortunately with my mediocre SQL that just made it horrendously slow. In particular if there were ghosts being checked with many links, it seems that all links would be pulled in. These could be for parts of the database that were not normally used, so it would result in a lot of I/O access and generate a large pause after tup was seemingly finished updating. For example, in the 'biggittup' project, creating a tmp.c file to include libbb.h, and then removing the include directive would result in the ghost check taking 141s with the old query. With the new queries, the same check takes 0.698s (this is on a cold cache).